colocate query generators with queries
