+++ apps/trading/pages/markets/__tests__/generate-market.ts
@@ -1,19 +1,7 @@
 import merge from 'lodash/merge';
 import type { PartialDeep } from 'type-fest';
+import type { Market } from '../__generated__/Market';
 
-interface Market_market {
-  __typename: 'Market';
-  id: string;
-  name: string;
-}
-
-interface Market {
-  market: Market_market | null;
-}
-
-export interface MarketVariables {
-  marketId: string;
-}
 export const generateMarket = (override?: PartialDeep<Market>): Market => {
   const defaultResult = {
     market: {
