+++ libs/chart/src/__tests__/generate-chart.ts
@@ -1,12 +1,12 @@
 import merge from 'lodash/merge';
 import type { PartialDeep } from 'type-fest';
-
-// TODO: Figure out how to get types
-// eslint-disable-next-line
-type Chart = any;
+import type {
+  Chart,
+  Chart_market_data_priceMonitoringBounds,
+} from '../lib/__generated__/Chart';
 
 export const generateChart = (override?: PartialDeep<Chart>): Chart => {
-  const priceMonitoringBound = {
+  const priceMonitoringBound: Chart_market_data_priceMonitoringBounds = {
     minValidPrice: '16256291',
     maxValidPrice: '18296869',
     referencePrice: '17247489',
